Add token-dashboard sagas to common
