+++ packages/mobile/src/screens/audio-screen/AudioScreen.tsx
@@ -201,7 +201,7 @@ export const AudioScreen = () => {
           {formatWei((totalBalance || new BN(0)) as BNWei, true, 0)}{' '}
         </Text>
         <View style={styles.audioInfo}>
-          {hasMultipleWallets ? (
+          {!hasMultipleWallets ? (
             <>
               <Text style={styles.audioText}>{messages.totalAudio}</Text>
               <TouchableOpacity
