+++ packages/web/src/services/env.ts
@@ -1,7 +1,8 @@
 export const env = {
   EAGER_DISCOVERY_NODES: process.env.REACT_APP_EAGER_DISCOVERY_NODES,
   EXPLORE_CONTENT_URL: process.env.REACT_APP_EXPLORE_CONTENT_URL,
-  SUGGESTED_FOLLOW_HANDLES: process.env.REACT_APP_SUGGESTED_FOLLOW_HANDLES,
+  SUGGESTED_FOLLOW_HANDLES: process.env
+    .REACT_APP_SUGGESTED_FOLLOW_HANDLES as string,
   AAO_ENDPOINT: process.env.REACT_APP_AAO_ENDPOINT,
   ORACLE_ETH_ADDRESSES: process.env.REACT_APP_ORACLE_ETH_ADDRESSES,
   ENVIRONMENT: process.env.REACT_APP_ENVIRONMENT as string
